+++ src/Control/Comonad/Env.hs
@@ -29,12 +29,7 @@ module Control.Comonad.Env
 -- * The EnvT comonad transformer
 , EnvT(..)
 , runEnvT
--- * Re-exported modules
-, module Control.Comonad
-, module Control.Comonad.Trans.Class
 ) where
 
-import Control.Comonad
 import Control.Comonad.Env.Class (ComonadEnv(..), asks)
-import Control.Comonad.Trans.Class
 import Control.Comonad.Trans.Env (Env, env, runEnv, EnvT(..), runEnvT, local)