+++ src/Control/Comonad/Identity.hs
@@ -1,19 +1,16 @@
-{-# LANGUAGE CPP #-}
 {-# LANGUAGE Safe #-}
 
 -- |
 -- Copyright   :  (C) 2008-2014 Edward Kmett
 -- License     :  BSD-style (see the file LICENSE)
---
 -- Maintainer  :  Edward Kmett <ekmett@gmail.com>
 -- Stability   :  experimental
 -- Portability :  non-portable (fundeps, MPTCs)
+
 module Control.Comonad.Identity
-( module Control.Comonad
-, module Data.Functor.Identity
+( module Data.Functor.Identity
 , module Control.Comonad.Trans.Identity
 ) where
 
-import Control.Comonad
 import Data.Functor.Identity
 import Control.Comonad.Trans.Identity