+++ src/Control/Comonad/Traced.hs
@@ -18,14 +18,7 @@ module Control.Comonad.Traced
 , runTraced
 -- * The TracedT comonad transformer
 , TracedT(..)
--- * Re-exported modules
-, module Control.Comonad
-, module Control.Comonad.Trans.Class
-, module Data.Monoid
 ) where
 
-import Control.Comonad
 import Control.Comonad.Traced.Class (ComonadTraced(..), traces)
-import Control.Comonad.Trans.Class
 import Control.Comonad.Trans.Traced (Traced, traced, runTraced, TracedT(..), runTracedT)
-import Data.Monoid